move SingleLabware under labware dir
